Add files via upload
The URL Shortener is a simple web-based application built using HTML, CSS, and JavaScript that allows users to convert long web links into short, easy-to-share URLs. The project focuses on enhancing user convenience by generating compact links, which can be copied and shared across social platforms, emails, or messages.

The frontend is designed with HTML for structure and CSS for styling to provide a clean and user-friendly interface. Users can paste any long URL into the input box, and with just one click, the system generates a shortened version. JavaScript is used to handle the core functionality, including validating the input, interacting with an API (like tinyurl or bitly), and dynamically displaying the shortened link on the same page without reloading.